+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## [1.5.13] - 2022-02-08
+- Additional fixes for Login & Authman sync [#106](https://github.com/rokwire/groups-building-block/issues/106)
+
+## [1.5.12] - 2022-02-07
+- Additional fixes for Login & Authman sync [#106](https://github.com/rokwire/groups-building-block/issues/106)
+
+## [1.5.11] - 2022-02-03
+-Additional fixes for Login & Authman sync [#106](https://github.com/rokwire/groups-building-block/issues/106)
 
 ## [1.5.10] - 2022-02-02
 - Add internal stats API [#112](https://github.com/rokwire/groups-building-block/issues/112)